Improve the precision of the intersection calculation: fewer false positives

+++ src/main/kotlin/io/github/antivanov/aoc2021/util/HashFunctions.kt
@@ -3,7 +3,7 @@ package io.github.antivanov.aoc2021.util
 object HashFunctions {
  private val PrimaryNumber = 37
 
- fun listOfNumbersHash(elements: List<Int>): Int  =
+ fun listOfNumbersHash(elements: List<Long>): Long  =
    elements.fold(1) { acc, it ->
      acc * PrimaryNumber + it
    }
